requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
cryptography==41.0.4
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-rest-auth==5.0.1; python_version >= '3.6'
django==4.2.6; python_version >= '3.8'
django-allauth==0.57.0; python_version >= '3.7'
djangorestframework==3.14.0; python_version >= '3.6'
drf-yasg==1.21.7; python_version >= '3.6'
idna==3.4; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pycparser==2.21
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python3-openid==3.2.0
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.8.0; python_version < '3.11'
tzdata==2023.3; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.0.6; python_version >= '3.7'

Django~=4.2.6
djangorestframework~=3.14.0